- Move `gi-pipeline` script into `gi-stats` - Disallow synthetic default imports - Make `dumpFile` synchronous
