+++ apps/frontend/src/app/Data/Weapons/WeaponSheet.tsx
@@ -4,7 +4,7 @@ import type {
   WeaponTypeKey,
 } from '@genshin-optimizer/consts'
 import { weaponAsset } from '@genshin-optimizer/gi-assets'
-import type { WeaponData } from '@genshin-optimizer/gi-pipeline'
+import type { WeaponData } from '@genshin-optimizer/gi-stats'
 import { getLevelString } from '@genshin-optimizer/gi-util'
 import { displayDataMap } from '.'
 import ImgIcon from '../../Components/Image/ImgIcon'
