+++ apps/frontend/tsconfig.json
@@ -6,13 +6,18 @@
     "noImplicitAny": false,
     "exactOptionalPropertyTypes": false,
     "strictNullChecks": false,
+    "allowSyntheticDefaultImports": true,
     "jsxImportSource": "@emotion/react"
   },
   "files": [],
   "include": [],
   "references": [
-    { "path": "./tsconfig.app.json" },
-    { "path": "./tsconfig.spec.json" }
+    {
+      "path": "./tsconfig.app.json"
+    },
+    {
+      "path": "./tsconfig.spec.json"
+    }
   ],
   "extends": "../../tsconfig.base.json"
 }
