+++ libs/gi-formula/src/data/char/util.ts
@@ -7,7 +7,7 @@ import type {
   WeaponTypeKey,
 } from '@genshin-optimizer/consts'
 import { locCharKeyToCharKey } from '@genshin-optimizer/consts'
-import type { CharacterDataGen } from '@genshin-optimizer/gi-pipeline'
+import type { CharacterDataGen } from '@genshin-optimizer/gi-stats'
 import type { NumNode } from '@genshin-optimizer/pando'
 import { prod, subscript, sum } from '@genshin-optimizer/pando'
 import type { Data, FormulaArg, Stat } from '../util'
