+++ libs/gi-formula/src/example.test.ts
@@ -20,7 +20,7 @@ import {
   userBuff,
 } from './data/util'
 import {} from './debug'
-import rawData from './example.test.json'
+import * as rawData from './example.test.json'
 import {
   artifactsData,
   charData,
