--- libs/gi-pipeline/.babelrc
@@ -1,10 +1,0 @@
-{
-  "presets": [
-    [
-      "@nx/web/babel",
-      {
-        "useBuiltIns": "usage"
-      }
-    ]
-  ]
-}
