--- libs/gi-pipeline/README.md
@@ -1,11 +1,0 @@
-# gi-pipeline
-
-This library was generated with [Nx](https://nx.dev).
-
-## Building
-
-Run `nx build gi-pipeline` to build the library.
-
-## Running unit tests
-
-Run `nx test gi-pipeline` to execute the unit tests via [Jest](https://jestjs.io).
