--- libs/gi-pipeline/project.json
@@ -1,31 +1,0 @@
-{
-  "name": "gi-pipeline",
-  "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "libs/gi-pipeline/src",
-  "projectType": "library",
-  "targets": {
-    "build": {},
-    "build-ts": {
-      "options": {
-        "outputFileName": "src/index.js"
-      }
-    },
-    "pipeline": {
-      "executor": "@nx/js:node",
-      "options": {
-        "buildTarget": "gi-pipeline:build-ts",
-        "watch": false,
-        "inspect": false
-      },
-      "inputs": ["default", "^production", "gi-stats"]
-    },
-    "lint": {}
-  },
-  "tags": [],
-  "namedInputs": {
-    "gi-stats": [
-      "{workspaceRoot}/lib/gi-stats/Data/**/*.json",
-      "{workspaceRoot}/lib/gi-stats/src/allStat_gen.json"
-    ]
-  }
-}
