+++ libs/gi-stats/package.json
@@ -1,5 +1,6 @@
 {
   "name": "@genshin-optimizer/gi-stats",
   "version": "0.0.1",
-  "type": "commonjs"
+  "type": "commonjs",
+  "executors": "./executors.json"
 }
