+++ libs/gi-stats/project.json
@@ -4,12 +4,16 @@
   "sourceRoot": "libs/gi-stats/src",
   "projectType": "library",
   "targets": {
-    "build": {},
-    "build-ts": {
-      "dependsOn": ["^build", "^pipeline"]
+    "genfile": {
+      "executor": "@genshin-optimizer/gi-stats:gen-stats",
+      "outputs": [
+        "{projectRoot}/Data/**/*",
+        "{projectRoot}/src/allStat_gen.json"
+      ]
     },
+    "build": {},
+    "build-ts": {},
     "lint": {}
   },
-  "tags": [],
-  "implicitDependencies": ["gi-pipeline"]
+  "tags": []
 }
