+++ libs/gi-stats/src/executors/gen-stats/schema.d.ts
@@ -1,0 +1,1 @@
+export interface GenStatsExecutorSchema {} // eslint-disable-line
