+++ libs/gi-stats/src/executors/gen-stats/src/QuantumCatalyst/data.json
