+++ libs/gi-stats/src/executors/gen-stats/src/Somnia/data.json
