+++ libs/gi-stats/src/executors/gen-stats/src/artifactData.ts
