+++ libs/gi-stats/src/executors/gen-stats/src/curves.ts
