+++ libs/gi-stats/src/executors/gen-stats/src/materialData.ts
