+++ libs/gi-stats/src/index.ts
@@ -1,6 +1,15 @@
 import type { CharacterKey } from '@genshin-optimizer/consts'
-import type { AllStats } from '@genshin-optimizer/gi-pipeline'
-import allStat_gen from './allStat_gen.json'
+import * as allStat_gen from './allStat_gen.json'
+
+// Make sure these are type-only imports/exports.
+// Importing the executor is quite costly.
+export type {
+  CharacterDataGen,
+  WeaponData,
+  WeaponDataGen,
+} from './executors/gen-stats/executor'
+import type { AllStats } from './executors/gen-stats/executor'
+
 const allStats = allStat_gen as AllStats
 
 export { allStats }
