+++ libs/sr-assets/tsconfig.json
@@ -2,11 +2,14 @@
   "compilerOptions": {
     "jsx": "react-jsx",
     "esModuleInterop": false,
-    "allowSyntheticDefaultImports": true,
     "types": ["vite/client"]
   },
   "files": [],
   "include": [],
-  "references": [{ "path": "./tsconfig.lib.json" }],
+  "references": [
+    {
+      "path": "./tsconfig.lib.json"
+    }
+  ],
   "extends": "../../tsconfig.base.json"
 }
