+++ libs/sr-stats/src/index.ts
@@ -1,5 +1,5 @@
 import type { AllStats } from '@genshin-optimizer/sr-pipeline'
-import allStat_gen from './allStat_gen.json'
+import * as allStat_gen from './allStat_gen.json'
 const allStats = allStat_gen as AllStats
 
 export { allStats }
