+++ libs/svgicons/tsconfig.json
@@ -2,12 +2,15 @@
   "compilerOptions": {
     "jsx": "react-jsx",
     "esModuleInterop": false,
-    "allowSyntheticDefaultImports": true,
     "jsxImportSource": "@emotion/react",
     "types": ["vite/client"]
   },
   "files": [],
   "include": [],
-  "references": [{ "path": "./tsconfig.lib.json" }],
+  "references": [
+    {
+      "path": "./tsconfig.lib.json"
+    }
+  ],
   "extends": "../../tsconfig.base.json"
 }
